# docs/api/openapi_cn.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentrade API",
    "version": "0.2.0",
    "description": "由 packages/contracts 生成的 v2 契约文档；无版本 API 请求会在运行时重定向到配置的默认版本。"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/v2/disputes": {
      "get": {
        "tags": [
          "Disputes"
        ],
        "operationId": "disputesListV2",
        "summary": "查询争议列表",
        "deprecated": false,
        "parameters": [
          {
            "in": "query",
            "name": "taskId",
            "required": false,
            "description": "任务 id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "opener",
            "required": false,
            "description": "发起方地址",
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "description": "争议状态筛选",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "RESOLVED_COMPLETED"
              ]
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "description": "按 id、发起方或争议双方说明搜索",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "排序字段",
            "schema": {
              "type": "string",
              "enum": [
                "latest",
                "created"
              ]
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "description": "排序方向",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "分页游标",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "分页大小",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDisputeResponse"
                }
              }
            }
          },
          "400": {
            "description": "输入或校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Disputes"
        ],
        "operationId": "disputesOpenV2",
        "summary": "发起争议",
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            }
          },
          "400": {
            "description": "输入或校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突或前置条件失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v2/disputes/{id}": {
      "get": {
        "tags": [
          "Disputes"
        ],
        "operationId": "disputesGetV2",
        "summary": "读取争议详情",
        "deprecated": false,
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "争议 id",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v2/disputes/{id}/votes": {
      "post": {
        "tags": [
          "Disputes"
        ],
        "operationId": "disputesVoteV2",
        "summary": "对争议投票",
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "争议 id",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteDisputeResult"
                }
              }
            }
          },
          "400": {
            "description": "输入或校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突或前置条件失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v2/disputes/{id}/counterparty-reason": {
      "post": {
        "tags": [
          "Disputes"
        ],
        "operationId": "disputesRespondV2",
        "summary": "提交争议对方说明",
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "争议 id",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            }
          },
          "400": {
            "description": "输入或校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突或前置条件失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/dashboard/summary": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "dashboardSummaryV2",
        "summary": "读取看板汇总",
        "deprecated": false,
        "parameters": [
          {
            "in": "query",
            "name": "tz",
            "required": false,
            "description": "IANA 时区",
            "schema": {
              "type": "string",
              "default": "UTC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "输入或校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminServiceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-service-key"
      }
    },
    "schemas": {
...
      "PaginatedDisputeResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dispute"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "VoteDisputeResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "vote",
          "workload"
        ],
        "properties": {
          "vote": {
            "$ref": "#/components/schemas/SupervisionVote"
          },
          "workload": {
            "$ref": "#/components/schemas/CycleWorkload"
          }
        }
      },
...
      "OpenDisputeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "taskId",
          "submissionId",
          "reasonMd"
        ],
        "properties": {
          "taskId": {
            "type": "string",
            "minLength": 1
          },
          "submissionId": {
            "type": "string",
            "minLength": 1
          },
          "reasonMd": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "VoteDisputeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "vote"
        ],
        "properties": {
          "vote": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "NOT_COMPLETED"
            ]
          }
        }
      },
      "RespondDisputeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reasonMd"
        ],
        "properties": {
          "reasonMd": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "V2ApiErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          }
        }
      }
    }
  }
}